Refactor LLM client and scriptwriter for improved JSON handling and shot management
- Enhance OllamaClient to strip markdown code fences from responses and handle empty or non-JSON responses gracefully.
- Update scriptwriter to enforce strict rules for video script generation, ensuring exactly 8 shots with proper scene prompts and narration.
- Introduce carry-over functionality for shots between episodes to manage excess shots effectively.
- Normalize shot durations to meet target duration requirements, ensuring consistency across episodes.
- Modify summarizer to enforce chronological order in key events and reduce arc summary length.
- Change ShotScript duration type from int to float for more precise timing.
- Improve VRAM management by using API calls for unloading models, enhancing performance and reliability.
- Add error handling in video assembly to ensure missing files are reported clearly.
- Implement ASS subtitle generation with karaoke effects for enhanced viewer engagement.

+++ models/schemas.py
@@ -9,7 +9,7 @@ class ShotScript(BaseModel):
 
     scene_prompt: str          # English, for ComfyUI
     narration_text: str        # Vietnamese, for TTS
-    duration_sec: int = 6
+    duration_sec: float = 6.0
     is_key_shot: bool = False
     characters: List[str] = []  # [] = scene-only shot, no IPAdapter
